% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crumbs-package.R
\docType{package}
\name{crumbs-package}
\alias{crumbs}
\alias{crumbs-package}
\title{crumbs: What the Package Does (Title Case)}
\description{
Framework for tracking changes to dataframes.
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

Authors:
\itemize{
  \item Jane Doe \email{jane@example.com}
}

}
\keyword{internal}
