% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_track.R
\name{descendants}
\alias{descendants}
\title{Select descendants of a stage}
\usage{
descendants(vars = NULL)
}
\arguments{
\item{vars}{A character vector of variable names. If not supplied,
the variables are taken from the current selection context (as
established by functions like \code{select()} or \code{pivot_longer()}).}
}
\description{
Select descendants of a stage
}
