% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_transformations.R
\name{edit_data}
\alias{edit_data}
\title{Apply edits to a dataframe}
\usage{
edit_data(data, edits, ...)
}
\arguments{
\item{data}{a stage or a dataframe}

\item{edits}{a list or dataframe of edits to be applied}

\item{...}{arguments passsed on to methods, primarily keys if data is a dataframe}

\item{na_alias}{an optional alias to use for NAs}
}
\description{
Apply edits to a dataframe
}
\examples{
print(5 + 5)
}
