% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{evolve}
\alias{evolve}
\title{Add new stages to a tree}
\usage{
evolve(.data, ..., .from)
}
\arguments{
\item{.data}{a tracked dataframe}

\item{...}{Name-value pairs of stages to add to the tree}

\item{.from}{Optional: vector of stages being used. If not provided, this
is automatically deduced from the expressions supplied.}
}
\description{
Add new stages to a tree
}
