% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_obj.R
\name{exclude}
\alias{exclude}
\title{Helper to exclude key combinations from a transformation}
\usage{
exclude(...)
}
\arguments{
\item{...}{list specifying key combinations to exclude}
}
\description{
Helper to exclude key combinations from a transformation
}
