% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{get_key_combs}
\alias{get_key_combs}
\title{Return all combinations of keys}
\usage{
get_key_combs(.data, ...)
}
\arguments{
\item{...}{stage names}

\item{tracked}{tracked dataframe}
}
\description{
Return all combinations of keys
}
