% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_tools.R
\name{history}
\alias{history}
\title{Compare snapshots}
\usage{
history(tree, values, ..., comp_names = "{.col}_{.name}", diffs = FALSE)
}
\arguments{
\item{tree}{a tracked_df object}

\item{values}{list of value columns}

\item{...}{snapshots to select}

\item{comp_names}{glue specification using {.col} and/or {.name}}

\item{diffs}{if TRUE, will display TRUE/FALSE}
}
\description{
Compare snapshots
}
