% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{merge_branches}
\alias{merge_branches}
\title{Combine snapshots}
\usage{
merge_branches(..., resolve = c("prefer_first", "prefer_last"))
}
\arguments{
\item{...}{snapshots to merge}

\item{resolve}{if prefer_first (the default), the first named snapshot will
be preferred when coalescing.}
}
\description{
Completes a full join of the data, then coalesces where there are columns with the same name.
}
\details{
NOT IN USE
}
