% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_obj.R
\name{print.exclude}
\alias{print.exclude}
\title{Print exclude object}
\usage{
\method{print}{exclude}(x, ...)
}
\arguments{
\item{x}{exclude object}

\item{...}{pass to method}
}
\description{
Print exclude object
}
