% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{print.snapshot_tree}
\alias{print.snapshot_tree}
\title{Print}
\usage{
\method{print}{snapshot_tree}(x, ...)
}
\arguments{
\item{x}{Object to format or print}

\item{...}{arguments to pass to methods}
}
\description{
Prints a snapshot tree shwoing the most recent frame
}
