% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{print.tracked_df}
\alias{print.tracked_df}
\title{Print}
\usage{
\method{print}{tracked_df}(x, ...)
}
\arguments{
\item{x}{Object to format or print}

\item{...}{arguments to pass to methods}
}
\description{
Prints a tracked DF showing the most recent stage
}
