% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quibble-package.R
\docType{package}
\name{quibble-package}
\alias{quibble}
\alias{quibble-package}
\title{quibble: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

Authors:
\itemize{
  \item Jane Doe \email{jane@example.com}
}

}
\keyword{internal}
