% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_tools.R
\name{replace_expr}
\alias{replace_expr}
\title{Replace names and symbols in an expression}
\usage{
replace_expr(expr, replace, replacement)
}
\arguments{
\item{expr}{Expression to modify}

\item{replace}{symbol/name to replace}

\item{replacement}{character replacement value}
}
\description{
Replace names and symbols in an expression
}
