% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_tools.R
\name{restore_keys}
\alias{restore_keys}
\title{Restore key combinatioins to a snapshot}
\usage{
restore_keys(data, to, which = "all")
}
\arguments{
\item{data}{snapshot}

\item{to}{snapshot which contains the key combinations you want to restore}

\item{which}{keys to restore - not in use yet}
}
\description{
One use case for this is restoring original keys
}
