% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_track.R
\name{select_stages}
\alias{select_stages}
\title{Select stage(s) using tidyselect-like syntax}
\usage{
select_stages(.data, ...)
}
\arguments{
\item{.data}{tracked dataframe}

\item{...}{Columns to select, unquoted, separated by commas. You can use
helpers as described here.}
}
\description{
Select stage(s) using tidyselect-like syntax
}
