% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{stage}
\alias{stage}
\title{Add new snapshots to a tree}
\usage{
stage(tree, ..., from)
}
\arguments{
\item{tree}{a snapshot tree}

\item{...}{Name-value pairs of snapshots to add to the tree}

\item{from}{Optional: the snapshot which is being used. If not provided, this
is automatically deduced from the expressions supplied.}
}
\description{
Add new snapshots to a tree
}
