% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track}
\alias{track}
\title{Track a dataframe through stages}
\usage{
track(..., keys)
}
\arguments{
\item{...}{dataframes (preferably one) to track}

\item{keys}{a vector of keys used to identify the data}
}
\description{
Track a dataframe through stages
}
\details{
This function creates a tracked_df object.
}
