% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_transformations.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate data}
\usage{
aggregate_data(data, exprs, col, value, exclude)
}
\arguments{
\item{data}{a snapshot}

\item{exprs}{list of expressions to add}

\item{col}{key column containing IDs to aggregate}

\item{value}{column containing values to aggregate}

\item{exclude}{an exclude() object - to be added}
}
\description{
The main additions compared to dplyr::mutate(!!!exprs) are
* Specify axis of aggregation
* Specify records to exclude
}
