% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{evolve}
\alias{evolve}
\title{Add new snapshots to a tree}
\usage{
evolve(.x, ..., .from)
}
\arguments{
\item{...}{Name-value pairs of snapshots to add to the tree}

\item{tree}{a snapshot tree}

\item{from}{Optional: the snapshot which is being used. If not provided, this
is automatically deduced from the expressions supplied.}
}
\description{
Add new snapshots to a tree
}
