% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{get_key_combs}
\alias{get_key_combs}
\title{Return all combinations of keys}
\usage{
get_key_combs(tracked, ...)
}
\arguments{
\item{tracked}{tracked dataframe}

\item{...}{snapshot names}
}
\description{
Return all combinations of keys
}
