% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{get_keys}
\alias{get_keys}
\title{Get keys from an object}
\usage{
get_keys(x, ...)
}
\arguments{
\item{x}{Either a snapshot or snapshot_tree}
}
\description{
This is a helper function which calls attr(x, "keys").
}
