% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_tools.R
\name{history_all}
\alias{history_all}
\title{Compare all snapshots within a tree
A special case of history()}
\usage{
history_all(tree, values, comp_names = "{.col}_{.name}")
}
\description{
Compare all snapshots within a tree
A special case of history()
}
