% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_tools.R
\name{history_all}
\alias{history_all}
\title{Compare all snapshots within a tree
This is a special case of history()}
\usage{
history_all(tree, values, comp_names = "{.col}_{.name}")
}
\arguments{
\item{tree}{a snapshot tree}

\item{values}{list of value columns}

\item{comp_names}{glue specification using {.col} and/or {.name}}
}
\description{
Compare all snapshots within a tree
This is a special case of history()
}
