% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_transformations.R
\name{overlay}
\alias{overlay}
\title{Overlay one dataframe onto another}
\usage{
overlay(x, y, by, ..., perl = TRUE)
}
\arguments{
\item{x}{dataframe to be edited}

\item{y}{dataframe containing edits}

\item{by}{vector of columns to join by}

\item{...}{These dots are for future extensions and must be empty.}

\item{perl}{use perl-compatible regex}
}
\description{
Overlay one dataframe onto another
}
