% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_tools.R
\name{plot.snapshot_tree}
\alias{plot.snapshot_tree}
\title{Plot snapshot tree as a graph}
\usage{
\method{plot}{snapshot_tree}(x, record = NULL, ...)
}
\arguments{
\item{x}{a snapshot tree}

\item{record}{optionally a named vector of key-value pairs, to show values for a specific key in the graph}

\item{...}{arguments to be passed to methods}
}
\description{
Plot snapshot tree as a graph
}
