% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_tools.R
\name{plot.tracked_df}
\alias{plot.tracked_df}
\title{Plot tracked dataframe as a graph}
\usage{
\method{plot}{tracked_df}(
  x,
  record = NULL,
  value_col = NULL,
  value_format = function(v) v,
  metric = c("value", "row_n"),
  plot_function,
  ...
)
}
\arguments{
\item{x}{a tracked_df object}

\item{record}{optionally a named vector of key-value pairs, to show values for a specific key in the graph}

\item{value_col}{name of a column to draw values from, if record is specified}

\item{value_format}{a function with single argument to format the value_col}

\item{metric}{metric to show: either value, row_n, or a summary function with argument table, which returns
a character of length 1 for each stage. Defaults to 'value' if record is provided}

\item{plot_function}{a function with arguement graph producing the output of the plot. can map
value, .in (int), .out (int), .type (normal/init/merge/final) and .label}

\item{...}{arguments to be passed to methods}
}
\description{
Plot tracked dataframe as a graph
}
