% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track}
\alias{track}
\title{Track a dataframe through stages}
\usage{
track(.data, keys, name = "init")
}
\arguments{
\item{keys}{a vector of keys used to identify the data}

\item{name}{name of the dataframe within the tracked_df}

\item{data}{dataframe to track}
}
\description{
Track a dataframe through stages
}
\details{
This function creates a tracked_df object.
}
