% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.R
\name{track}
\alias{track}
\title{Track a dataframe through stages}
\usage{
track(..., keys)
}
\arguments{
\item{...}{dataframes (preferably one) to track}

\item{keys}{a vector of keys used to identify the data}
}
\description{
Track a dataframe through stages
}
\details{
This function returns an object of class snapshot_tree, which is a list containing
 keyed data (snapshots) which have been evolve()d, as well as the parents of each snapshot.
}
